Solution to Calculate and Print the Average of Three Numbers

+++ Calculate and Print the Average of Three Numbers@@6788d5012db25efec946f944/code.c
@@ -1,8 +1,8 @@
 #include <stdio.h>
 int main() {
-    float a,b;
+    int a,b;
     float c;
-    scanf("%f %f",&a &b);
+    scanf("%d %d",&a &b);
     c=(a*b)/2;
     printf("Average: %f\n",c);
     return 0;
